+++ src/game/client/components/scoreboard.h
@@ -33,7 +33,7 @@ class CScoreboard : public CComponent
 
 	void RenderGoals(float Width, float Height);
 	void RenderSpectators(float Width, float y);
-	int RenderScoreboard(float Width, float y, int Team, const char *pTitle, bool TeamPlay);
+	float RenderScoreboard(float Width, float y, int Team, const char *pTitle, bool TeamPlay);
 	void RenderRecordingNotification(float x);
 
 	static void ConKeyScoreboard(IConsole::IResult *pResult, void *pUserData);
